LockGuide if first time
* EmotionExercise.EActivity to flag to check if instruction readed
* Player: save activities instructions read
* PopUp: call onclose on Close function
* Guide PopUp:
 + Lock play if Player never read guide
 + Guide animation
 + Avatar generic icon

+++ Assets/Scripts/Emotions/EmotionToggles.cs
@@ -124,7 +124,7 @@ public class EmotionToggles : MonoBehaviour
         if(gm.IsPlayerActive())
         {
             Emotion.EEmotions emotionsChecked = GetEmotionsFlagSelected();
-            gm.currentPlayer.EmotionsLearned = (uint)emotionsChecked;
+            gm.currentPlayer.EmotionsLearned = emotionsChecked;
             Debug.Log("EmotionToggles: SaveEmotionsChecked: " + emotionsChecked);
             if (numActive > 0)
                 DBManager.Instance.UpdatePlayerToDb(gm.currentPlayer);
